Mobile UI updates / Complaint Hearing Dates

+++ database/migrations/2021_09_13_100551_create_outside_complainants_table.php
@@ -17,7 +17,7 @@ class CreateOutsideComplainantsTable extends Migration
             $table->unsignedInteger('compId');
             $table->string('complainant');
             $table->string('address');
-            $table->integer('contact');
+            $table->bigInteger('contact');
             $table->timestamps();
             $table->foreign('compId')
             ->references('id')
